remove all omni lock version test cases.
