--- integration-test/src/tests_omni_lock/mod.rs
@@ -1,3 +1,0 @@
-mod helper;
-mod payment;
-mod rpc;
