--- integration-test/src/tests_omni_lock/payment/mod.rs
@@ -1,1 +1,0 @@
-mod dust_collector;
